+++ refinedstorage-jei-integration-neoforge/src/main/java/com/refinedmods/refinedstorage/jei/forge/ForgePlatform.java
@@ -1,13 +1,13 @@
 package com.refinedmods.refinedstorage.jei.forge;
 
+import com.refinedmods.refinedstorage.common.support.resource.FluidResource;
 import com.refinedmods.refinedstorage.jei.common.Platform;
-import com.refinedmods.refinedstorage.platform.common.support.resource.FluidResource;
 
 import java.util.Optional;
 
 import net.neoforged.neoforge.fluids.FluidStack;
 
-import static com.refinedmods.refinedstorage.platform.neoforge.support.resource.VariantUtil.ofFluidStack;
+import static com.refinedmods.refinedstorage.neoforge.support.resource.VariantUtil.ofFluidStack;
 
 public class ForgePlatform implements Platform {
     @Override